Segment_model/E2EPTB-codes/src/CUnet/requirements.txt
https://download.pytorch.org/whl/cu90/torch-1.1.0-cp36-cp36m-linux_x86_64.whl
tensorboardX
torchvision
torchvis
torch
numpy
sklearn
Augmentor
Pillow
pandas
opencv-python
